ENH: added support for the noquote tag in bmms.
This is not a very robust implementation, more like useful for a one-off
situation.  It could be the basis for a better implementation, if thought
over more carefully.  What this will do, is if a parameter has the
noquote tag in the bmm, that parameter will not be quoted in the condor
script that gets written out by BatchMake.